move to TS; change readme;
