+++ .prettierrc
@@ -1,0 +1,11 @@
+{
+  "trailingComma": "es5",
+  "tabWidth": 2,
+  "semi": true,
+  "singleQuote": true,
+  "bracketSpacing": true,
+  "printWidth": 90,
+  "arrowParens": "avoid",
+  "parser": "typescript",
+  "endOfLine": "lf"
+}
