+++ README.md
@@ -10,6 +10,8 @@
 **Development**
 
 `npm run start:dev`
+or
+`npm run dev`
 
 * App served @ `http://localhost:8181` with nodemon
 
@@ -26,6 +28,7 @@
 Command | Description
 --- | ---
 `npm run start:dev` | App served @ `http://localhost:8181` with nodemon
+`npm run dev` | App served @ `http://localhost:8181` with nodemon
 `npm run start` | App served @ `http://localhost:8181` without nodemon
 
 **Note**: replace `npm` with `yarn` in `package.json` if you use yarn.
