+++ package.json
@@ -4,16 +4,27 @@
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
     "type": "module",
+    "exports": "./dist/index.js",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "webpack && node ./dist/bundle.js",
+        "start:dev": "nodemon",
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "dev": "npm run start:dev"
     },
-    "author": "Andrei Auchynnikau",
+    "author": "4ebula",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/node": "^20.3.2",
+        "@typescript-eslint/eslint-plugin": "^5.60.1",
+        "@typescript-eslint/parser": "^5.60.1",
+        "eslint": "^8.43.0",
+        "nodemon": "^2.0.16",
+        "ts-loader": "^9.4.4",
+        "ts-node": "^10.9.1",
+        "typescript": "^5.1.6",
+        "webpack": "^5.88.1",
+        "webpack-cli": "^5.1.4"
     },
     "dependencies": {
         "ws": "^8.8.0"
