+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+  "compilerOptions": {
+    "outDir": "./dist",
+    "noImplicitAny": true,
+    "module": "ES2020",
+    "target": "ES2020",
+    "allowJs": true,
+    "moduleResolution": "node",
+    "esModuleInterop": true,
+    "baseUrl": ".",
+    "rootDir": "./",
+    "lib": ["ES2020"],
+    "removeComments": true,
+    "allowSyntheticDefaultImports": true,
+    "forceConsistentCasingInFileNames": true,
+  },
+  "ts-node": {
+      "esm": true,
+      "experimentalSpecifierResolution": "node"
+  },
+  "exclude": ["node_modules", "front"],
+  "include": ["src/**/*.ts"],
+}
