+++ webpack.config.js
@@ -1,0 +1,35 @@
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+export default {
+  entry: path.resolve(__dirname, './src/index.ts'),
+  target: 'node',
+  module: {
+    rules: [
+      {
+        test: /\.ts?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
+  plugins: [
+    new Dotenv()
+  ],
+  resolve: {
+    extensions: ['.ts', '.js', '.json'],
+  },
+  output: {
+    filename: 'bundle.js',
+    path: path.resolve(__dirname, 'dist'),
+    library: {
+      type: "module",
+    },
+    chunkFormat: 'module',
+  },
+  experiments: {
+    outputModule: true,
+  },
+};
